Integrate hypothesis panel
